+++ src/index.js
@@ -3,40 +3,16 @@ const cors = require("@fastify/cors");
 const swagger = require("@fastify/swagger");
 const swaggerUI = require("@fastify/swagger-ui");
 const eligibilityService = require("./services/eligibilityService");
-const validationService = require("./services/validationService");
+// const validationService = require("./services/validationService");
 const swaggerConfig = require("./config/swagger");
-
+const benefitEligibleSchema = require('./schemas/check-eligibility-schema');
+const userEligibilitySchema = require('./schemas/check-users-eligibility-schema')
 // Register plugins
 fastify.register(cors, {
   origin: "*",
   methods: ["GET", "POST", "PUT", "DELETE"],
 });
 
-// Define schemas
-const userProfileSchema = {
-  type: "object",
-  required: ["name", "gender", "age", "dateOfBirth", "caste", "income"],
-  properties: {
-    name: { type: "string", description: "Full name of the user" },
-    gender: {
-      type: "string",
-      description: "Gender of the user",
-      enum: ["male", "female"],
-    },
-    age: { type: "number", description: "Age of the user" },
-    dateOfBirth: {
-      type: "string",
-      format: "date",
-      description: "Date of birth in YYYY-MM-DD format",
-    },
-    caste: {
-      type: "string",
-      description: "Caste category",
-      enum: ["sc", "st", "obc", "general"],
-    },
-    income: { type: "number", description: "Annual income in INR" },
-  },
-};
 
 // Register Swagger
 fastify.register(swagger, {
@@ -106,138 +82,7 @@ fastify.setErrorHandler((error, request, reply) => {
 fastify.post(
   "/check-eligibility",
   {
-    schema: {
-      tags: ["Eligibility"],
-      summary: "Check eligibility for benefits",
-      description:
-        "Checks if a user is eligible for all available benefit schemes",
-      body: {
-        type: "object",
-        required: ["userProfile", "benefitSchemas", "customRules"],
-        properties: {
-          userProfile: userProfileSchema,
-          benefitSchemas: {
-            type: "array",
-            items: {
-              type: "object",
-              required: ["en"],
-              properties: {
-                en: {
-                  type: "object",
-                  required: ["eligibility"],
-                  properties: {
-                    eligibility: {
-                      type: "array",
-                      items: {
-                        type: "object",
-                        required: ["type", "description", "criteria"],
-                        properties: {
-                          type: { type: "string" },
-                          description: { type: "string" },
-                          criteria: {
-                            type: "object",
-                            required: ["name", "condition", "conditionValues"],
-                            properties: {
-                              name: { type: "string" },
-                              condition: { type: "string" },
-                              conditionValues: {
-                                type: ["string", "number", "array"],
-                              },
-                            },
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-          customRules: {
-            type: "object",
-            additionalProperties: true,
-            description: "Custom rules to override standard eligibility checks",
-          },
-        },
-      },
-      response: {
-        200: {
-          type: "object",
-          properties: {
-            eligible: {
-              type: "array",
-              items: {
-                type: "object",
-                properties: {
-                  schemaId: {
-                    type: "string",
-                    description: "ID of the eligible benefit scheme",
-                  },
-                },
-              },
-            },
-            ineligible: {
-              type: "array",
-              items: {
-                type: "object",
-                properties: {
-                  schemaId: {
-                    type: "string",
-                    description: "ID of the ineligible benefit scheme",
-                  },
-
-                  reasons: {
-                    type: "array",
-                    items: {
-                      type: "object",
-                      properties: {
-                        type: { type: "string" },
-                        field: { type: "string" },
-                        reason: { type: "string" },
-                        description: { type: "string" },
-                        userValue: { type: ["string", "number"] },
-                        requiredValue: { type: ["string", "number", "array"] },
-                        condition: { type: "string" },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-            errors: {
-              type: "array",
-              items: {
-                type: "object",
-                properties: {
-                  schemaId: { type: "string" },
-                  error: { type: "string" },
-                },
-              },
-            },
-          },
-        },
-        400: {
-          type: "object",
-          properties: {
-            error: {
-              type: "string",
-              description: "Error message",
-            },
-            message: { type: "string", description: "Detailed error message" },
-          },
-        },
-        500: {
-          type: "object",
-          properties: {
-            error: {
-              type: "string",
-              description: "Error message",
-            },
-            message: { type: "string", description: "Detailed error message" },
-          },
-        },
-      },
-    },
+    schema: benefitEligibleSchema
   },
   async (request, reply) => {
     try {
@@ -261,182 +106,7 @@ fastify.post(
 fastify.post(
   "/check-users-eligibility",
   {
-    schema: {
-      tags: ["Eligibility"],
-      summary: "Check eligibility of multiple users for a benefit scheme",
-      description:
-        "Checks if multiple users are eligible for a specific benefit scheme and returns detailed reasons for ineligibility",
-      body: {
-        type: "object",
-        required: ["userProfiles", "benefitSchema"],
-        properties: {
-          userProfiles: {
-            type: "array",
-            items: {
-              type: "object",
-              required: [
-                "name",
-                "gender",
-                "age",
-                "dateOfBirth",
-                "caste",
-                "income",
-              ],
-              properties: {
-                name: { type: "string", description: "Full name of the user" },
-                gender: {
-                  type: "string",
-                  enum: ["male", "female"],
-                  description: "Gender of the user",
-                },
-                age: { type: "number", description: "Age of the user" },
-                dateOfBirth: {
-                  type: "string",
-                  format: "date",
-                  description: "Date of birth in YYYY-MM-DD format",
-                },
-                caste: {
-                  type: "string",
-                  enum: ["sc", "st", "obc", "general"],
-                  description: "Caste category",
-                },
-                income: { type: "number", description: "Annual income in INR" },
-                documents: {
-                  type: "object",
-                  description: "User documents with verification status",
-                  additionalProperties: {
-                    type: "object",
-                    properties: {
-                      verified: {
-                        type: "boolean",
-                        description: "Whether the document is verified",
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-          benefitSchema: {
-            type: "object",
-            required: ["en"],
-            properties: {
-              en: {
-                type: "object",
-                required: ["eligibility"],
-                properties: {
-                  eligibility: {
-                    type: "array",
-                    items: {
-                      type: "object",
-                      required: ["type", "description", "criteria"],
-                      properties: {
-                        type: { type: "string" },
-                        description: { type: "string" },
-                        criteria: {
-                          type: "object",
-                          required: ["name", "condition", "conditionValues"],
-                          properties: {
-                            name: { type: "string" },
-                            condition: { type: "string" },
-                            conditionValues: {
-                              anyOf: [
-                                { type: "string" },
-                                { type: "number" },
-                                {
-                                  type: "array",
-                                  items: {
-                                    anyOf: [
-                                      { type: "string" },
-                                      { type: "number" },
-                                    ],
-                                  },
-                                  minItems: 1,
-                                },
-                              ],
-                            },
-                          },
-                        },
-                        allowedProofs: {
-                          type: "array",
-                          items: { type: "string" },
-                          description:
-                            "List of allowed document types for verification",
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      },
-      response: {
-        200: {
-          type: "object",
-          properties: {
-            eligibleUsers: {
-              type: "array",
-              items: {
-                type: "object",
-                properties: {
-                  name: {
-                    type: "string",
-                    description: "Name of the eligible user",
-                  },
-                  applicationId: {
-                    type: "string",
-                    description: "Application ID of the eligible user",
-                  },
-                },
-              },
-            },
-            ineligibleUsers: {
-              type: "array",
-              items: {
-                type: "object",
-                properties: {
-                  name: {
-                    type: "string",
-                    description: "Name of the ineligible user",
-                  },
-                  applicationId: {
-                    type: "string",
-                    description: "Application ID of the eligible user",
-                  },
-                  reasons: {
-                    type: "array",
-                    items: { type: "string" },
-                    description: "List of reasons why the user is ineligible",
-                  },
-                },
-              },
-            },
-          },
-        },
-        400: {
-          type: "object",
-          properties: {
-            error: {
-              type: "string",
-              description: "Error message",
-            },
-            message: { type: "string", description: "Detailed error message" },
-          },
-        },
-        500: {
-          type: "object",
-          properties: {
-            error: {
-              type: "string",
-              description: "Error message",
-            },
-            message: { type: "string", description: "Detailed error message" },
-          },
-        },
-      },
-    },
+    schema: userEligibilitySchema
   },
   async (request, reply) => {
     try {
